Subtype from generic Formatter so making new formatters is easier
