+++ rsmf/setup.py
@@ -6,10 +6,10 @@ import re
 import sys
 from pathlib import Path
 
-from .quantumarticle import parse as quantumarticle_parse
-from .revtex import parse as revtex_parse
+from .quantumarticle import QuantumarticleFormatter
+from .revtex import RevtexFormatter
 
-_PARSERS = [quantumarticle_parse, revtex_parse]
+_FORMATTERS = [QuantumarticleFormatter, RevtexFormatter]
 
 _comment_regex = re.compile("(%.*)")
 
@@ -50,6 +50,35 @@ def _extract_preamble(path):
     return "".join(lines)
 
 
+def _parse(preamble):
+    """Parse the preamble to extract the informations relevant for plot style.
+
+    Args:
+        preamble (str): The preamble, containing at least the \documentclass command.
+
+    Returns:
+        Dict: kwargs for the Formatter (columns, paper, fontsize)
+    """
+    kwargs = {}
+
+    if "onecolumn" in preamble:
+        kwargs["columns"] = "onecolumn"
+    else:
+        kwargs["columns"] = "twocolumn"
+
+    if "letterpaper" in preamble:
+        kwargs["paper"] = "letterpaper"
+    else:
+        kwargs["paper"] = "a4paper"
+
+    if "11pt" in preamble:
+        kwargs["fontsize"] = 11
+    elif "12pt" in preamble:
+        kwargs["fontsize"] = 12
+    else:
+        kwargs["fontsize"] = 10
+
+
 def setup(arg):
     """Get a formatter corresponding to the document's class.
 
@@ -68,17 +97,13 @@ def setup(arg):
         preamble = arg
 
     preamble = _clean_preamble(preamble)
+    formatter_kwargs = _parse(preamble)
 
     result = None
-    for parser in _PARSERS:
-        result = parser(preamble)
+    for CustomFormatter in _FORMATTERS:
+        if CustomFormatter._documentclass_identifier in preamble:
+            return CustomFormatter(**kwargs)
 
-        if result:
-            break
-
-    if result is None:
-        raise Exception(
+    raise Exception(
             "No formatter was found for the given argument. This means either there is no formatter, or, if you gave a file path that it does not exist."
         )
-
-    return result
